requirements.txt
anyio==3.7.0
click==8.1.3
exceptiongroup==1.1.2
fastapi==0.99.1
h11==0.14.0
idna==3.4
pydantic==1.10.11
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
sqlalchemy==1.4.25
python-decouple==3.4